+++ demo/App.tsx
@@ -11,6 +11,7 @@ import './App.css';
 import './assets/css/App.css';
 import './assets/_scss/style.scss';
 import { cssModule } from './style.module.css';
+import { scssModule } from './style.module.scss';
 
 function App() {
   const [count, setCount] = useState(0);
@@ -20,6 +21,7 @@ function App() {
         <img src="/logo.svg" className="App-logo" alt="logo" />
         <img src={logo2} className="logo2" alt="logo2" />
         <p>Hello Vite + React!</p>
+        <p className={scssModule}>Styled by SCSS Modules</p>
         <StyledText>This text is styled by styled-components</StyledText>
         <p className="global-css">
           This text is styled by global css which is not imported to App.tsx
@@ -29,6 +31,9 @@ function App() {
           This text is styled by global configured SASS
         </p>
         <p className="imported-sass">This text is styled by imported SASS</p>
+        <p className="bg-yellow-400 font-bold m-5 text-red-500">
+          This text is styled by Tailwind CSS
+        </p>
         <button
           css={css`
             padding: 32px;